+++ docs/develop/change-log.md
@@ -1,4 +1,8 @@
 # Change Log
+### Version 2.2.2
+
+- [x] config: (break change) `http_redirect_to_https` convert from bool to u32.
+- [x] chore: change api domain response type.
 
 ### Version 2.2.1